requirements.txt
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
shap>=0.44.0
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.32.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
plotly>=5.18.0
python-dotenv>=1.0.0
httpx>=0.26.0
